Solutions to day 7 and 8
Both a little more tricky. In day 7 an understanding of median, mean and the sum of an arithmetic series is handy.
Day 8 took a ling time to get the part 2 logic correct. From the point of view of Ada code look at the overloading of the Boolean operators and, or, and  not.