Day 12 Solution
In part two Find_Paths  is run multiple times with one specific cave allowed to be visited twice. There is an interesting question, is this equivalent to running something similar to Find_Paths but doing a check at each invocation to verify that not more than one small cave has been visited twice?