Solution to Day 14
I needed help for this one, I got a clue from reddit. The secret is to consider populations rather than attempting to generate the polymer. The latter can be don for part one but is infeasible for part two. In concept the solution is similar to day six Lantern Fish problem.
The other tricky bits are the two "edge cases" and the double counting of letter pairs.