Update december_16.adb
Tidy up removed a redundant declaration and corrected a comment.
One issue I missed on the first push was that the result has to be represented as an unsigned_64. Initially I used unsigned 32 and got an answer that was too small. in Ada the Unsigned_xx types are modular so there cannot be an overflow exception. The Unsigned_xx types are the only types that bit wise logical operations can be applied to.