Day 18 Solution
You had to read the instructions relating to explode and split carefully. The key words being regular numbers as opposed to digits! It was easy to assume that because there were no examples of an explode where one of the numbers contained two digits before the explode that a split was required immediately. Allowing for two or more digit numbers was a complication, but resulted in cleaner code. Using Slice to copy the remainder of the string after the explode reduced the execution time from 3.6 s to 2.4 s. Once the reduction process was correct the rest was reasonably straight forward. In part two I did the additions both ways, that is A + B and B + A, this is probably necessary because of the multiply by two for the left and three for the right almost certainly means that the addition is not commutative. That is A + B /= B + A for Snailfish numbers!