Day 22 Part 1 Solution
Based on the size of the numbers in part 2 it is impractical to store the state of individual cubed. A strategy that may work is to record the on cubes as their bounding coordinates as per input. When partial overlaps these can be managed by splitting the partially intersected rectangular into a series of smaller rectangular prisms. That way only intersections of rectangular prisms have to be managed.