Part two solved
My solution for part two reduces the problem to a single dimension performing calculations on lines parallel to the X axis. I got this idea from a megatrend post but initially discarded the idea as it implied iterating over Z, Y and the all prisms, For my input of the order of 1.3*10**13 lines! The elegant part is only calculating lines that change as a result of a Z or Y boundary of a prism. This reduces the number of lines to be calculated to something less than 4,000,000. This got the execution time for the example down to 300ms and 7.5s for my input.

  My test cases were intended to verify calculations for rectangles in a plane. I could not get this to work. the test cases are all in the -50..50 range and thus yield the same result for parts one and two.