Day 24 Solution to both parts
Really tricky, the solution draws in the push and pop concepts that are a recurring theme on the Day 24 Megatrend. The game has been played in that there is a functional virtual machine that is used to find the solution and also to verify that the is potentially valid in that it returns z = 0.