Day 24 part two complete, AOC 2021 complete!
My biggest problem with Day 24 was the concept of universes. Until I looked at the Megatrend I did not appreciate that this equated to all possible outcomes (unique or otherwise). The one point I missed initially was tracking possible outcomes up to the a player wins. The solution here uses a queue to sore the state of all uncompleted game instances. No cashing is used and the run time for part two is about 10 s for the example. Interestingly the maximum number of game instances in the queue is 3946912 for the example. I got the queue idea from Megatrends; however when I went back I could not find the reference, I believe this is a reasonably elegant solution. The solution is now self contained, functionality of the deterministic die has been incorporated into the Next_Position procedure on part one.